search pills fixup: Group pill support, except closing users.
This supports adding users to DM pills by simply typing
a user's name in the text input directly after a complete
dm pill. It only works for DM pills at the end of search
input.

This commit doesn't include functionality for removing
users from a container pill, which should eventually
be squashed with this commit, but is currently separate
because I'm pretty unsure about the implementation
and wanted to isolate that code for easier review.