search pills fixup: Add ability to remove users from user pills.
Clicking the X button removes a user from its user pill
container, and if that user was the last user left in the
container, then the whole container is also removed.

This commit should eventually be squashed into the commit
with group user pills, but the implementation feels
maybe wrong so I separated it out for easier review.