typeahead: Add hideAfterSelect option.
See #26803 for more context on when this was introduced for
search pills. It was needed for "search for..." terms (i.e.
terms with operator "search"), which we'll be displaying
as text input, instead of pillifying it. Because we leave
it as text input, the text input field could end up not
empty, which prevents us from closing the typeahead after
selection, even though `helpOnEmptyStrings` is `false` for
the search typeahead.

We need to treat this situation like there's an empty string,
so we use `hideAfterSelect` to force this.